requirements.txt
redis==5.0.7
psycopg2-binary==2.9.9
requests==2.32.3
prometheus-client==0.20.0
